bilibili-oneclick/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
bilibili-api-python
